quotient/requirements.txt
aerich==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
aiocache[msgpack]==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp-asgi==0.5.2 ; python_version >= "3.11" and python_version < "4.0"
aiohttp>=3.9.0 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.0 ; python_version >= "3.11" and python_version < "4.0"
astunparse==1.6.3 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.11" and python_version < "4.0"
asyncpg>=0.29.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
bidict==0.22.1 ; python_version >= "3.11" and python_version < "4.0"
braceexpand==0.1.7 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
dateparser==1.1.8 ; python_version >= "3.11" and python_version < "4.0"
dictdiffer==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
discord-py==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.98.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist>=1.4.1 ; python_version >= "3.11" and python_version < "4.0"
humanize==4.6.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
imagehash==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
import-expression==1.1.4 ; python_version >= "3.11" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
jishaku==2.5.1 ; python_version >= "3.11" and python_version < "4.0"
# line-profiler==4.0.3 ; python_version >= "3.11" and python_version < "4.0"
lru-dict==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
multidict>=6.0.5 ; python_version >= "3.11" and python_version < "4.0"
numpy>=1.26.0 ; python_version >= "3.11" and python_version < "4.0"
parsedatetime==2.6 ; python_version >= "3.11" and python_version < "4.0"
pillow>=10.2.0 ; python_version >= "3.11" and python_version < "4.0"
prettytable==3.8.0 ; python_version >= "3.11" and python_version < "4.0"
psutil==5.9.5 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.9 ; python_version >= "3.11" and python_version < "4.0"
# pygit2>=1.13.0 ; python_version >= "3.11" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-engineio==4.4.1 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.11" and python_version < "4.0"
python-socketio==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
pywavelets>=1.5.0 ; python_version >= "3.11" and python_version < "4.0"
regex==2023.6.3 ; python_version >= "3.11" and python_version < "4.0"
scipy>=1.12.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.11.8 ; python_version >= "3.11" and python_version < "4.0"
tortoise-orm==0.19.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
tzlocal==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
ujson==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.11" and python_version < "4.0"
wheel==0.40.0 ; python_version >= "3.11" and python_version < "4.0"
yarl>=1.9.4 ; python_version >= "3.11" and python_version < "4.0"
